requirements.txt
# ---- API & Web ----
fastapi
uvicorn
twilio

# ---- Data ----
pandas

# ---- AI / RAG ----
langchain
langchain-community
langchain-google-genai
google-generativeai
sentence-transformers

# ---- Utils ----
python-dotenv
pydantic
